feat(create new component): create new component CreateVacancy
