+++ src/components/__test__/CreateVacancy.test.js
@@ -1,0 +1,8 @@
+import renderer from "react-test-renderer";
+
+import CreateVacancy from "../CreateVacancy";
+
+it("renders correctly", () => {
+    const tree = renderer.create(<CreateVacancy />).toJSON();
+    expect(tree).toMatchSnapshot();
+});
