+++ src/pages/careers/index.jsx
@@ -4,15 +4,9 @@ import AvailableJobs from "@/components/AvailableJobs";
 import CareerPhilosophy from "@/components/CareerPhilosophy";
 import getDocument from "@/firebase/getData";
 import PageIntro from "@/components/PageIntro";
-import Button from "@/components/ui/Button";
-import Router from "next/router";
+import CreateVacancy from "@/components/CreateVacancy";
+
 function career({ t, careers }) {
-    const handleClick = () => {
-        const router = require("next/router").default;
-        router.push({
-            pathname: "/careers/create",
-        });
-    };
     return (
         <div>
             <div className=' lg:ms-28 mt-10 ms-10 '>
@@ -28,21 +22,7 @@ function career({ t, careers }) {
                 <AvailableJobs t={t} careers={careers} />
             </div>
 
-            <div className='container pb-8'>
-                <PageIntro
-                    title={t("doYouHavevacancy")}
-                    subtitle={t("doYouHavevacancyDescrption")}
-                />
-                <Button
-                    content={t("addToCareersList")}
-                    textTransform='uppercase'
-                    filled='true'
-                    size='large'
-                    fontSize='text-lg md:text-xl lg:text-2xl'
-                    radius='md'
-                    onClick={handleClick}
-                />
-            </div>
+            <CreateVacancy t={t} />
         </div>
     );
 }
